requirements.txt
pyspark==3.5.6
confluent-kafka==2.11.0
python-dotenv==1.1.1
